Add files via upload
This project was done in completion of a Product management Nanodegree.
There are 4 files in this directory representing the four parts of the project. For the first part, I Pitched a product vision for a product, “FITPAL”, which was designed to help members of Kaiser Permenante, a healthcare service provider, to lead a healthier life devoid of non- communicable diseases that develop as a result of inactivity.

For the second part of this project, we ran a design sprint where we developed the concept of this product that addresses the opportunity we identified.

For the third part, we carried out all of the activities that product managers commit to during product development. We worked on identifying risks by creating a test plan and getting feedback from dogfooders. This feedback was then prioritized and a go/no go meeting was simulated.

For the fourth and final part of this project, we simulated a product launch, where we brought the product to the potential market. As a product manager, it’s important to collaborate with various teams to ensure a smooth and successful product launch. Post launch, the feedback from users was used to determine the next steps for the product.